chore: update npm packages within semver limits
The update included a new version of Prisma 1.31 which includes
a new Prisma Datamodel (1.1), thus there were some changes to
the Prisma docker configuration to address this and the
datamodel.prisma file. Notably, `prisma deploy` now automatically
runs `prisma generate` so I removed the previous post-deploy hook
which did this manually as it is redundant now.

Signed-off-by: David Piper <david.piper@mailfence.com>

ci: fix ci/prisma which was broken by requesting too many db connections

The cause of the broken CI was that Prisma was trying
to make more connections that Postgres allows (100 by default;
Prisma wanted 200). So I fixed this by having Prisma only
request 50. Then the data pulling code will now work
in batches of 15 instead of 50 as it was originally.

I am not sure exactly what changed, but I suspect that
it has something to do with an undocumented change in Prisma
1.31.

---------------------------------------------------------------------

ci: remove db connection limit; it seems to be breaking ci

Signed-off-by: David Piper <david.piper@mailfence.com>

ci: use smaller batch size for adding data

The amount of data added to the DB at once might be
too much and overwhelming it, lowering it significantly.
Also remove the redundant @types/is-reachable since
is-reachable now provides its own TypeScript type
definitions.

Signed-off-by: David Piper <david.piper@mailfence.com>

chore(server): add missing newline to datamodel; found during review

Signed-off-by: David Piper <david.piper@mailfence.com>

ci: put db limit back; lower requests in data adding

The cause of the broken CI was that Prisma was trying
to make more connections that Postgres allows (100 by default;
Prisma wanted 200). So I fixed this by having Prisma only
request 50. Then the data pulling code will now work
in batches of 15 instead of 50 as it was originally.

I am not sure exactly what changed, but I suspect that
it has something to do with an undocumented change in Prisma
1.31.

Signed-off-by: David Piper <david.piper@mailfence.com>

---------------------------------------------------------------------

Signed-off-by: David Piper <david.piper@mailfence.com>